# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ChatCommunicator REST API docs",
    "version": "in dev"
  },
  "paths": {
...
    "/api/friend-invitations": {
      "post": {
        "tags": [
          "FriendInvitations"
        ],
        "summary": "Send friendship invitation",
        "description": "Route: POST api/friend-invitations\r\nAuthorization: Required.\r\nDuplicate or conflicting invitations are handled by domain logic.",
        "requestBody": {
          "description": "The request payload containing the sender and recipient identifiers.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFriendInvitationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFriendInvitationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SendFriendInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitation created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendFriendInvitationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload or business rule violation."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "Sender not authorized to perform this action."
          },
          "404": {
            "description": "Sender or recipient user not found."
          },
          "409": {
            "description": "An active invitation or friendship already exists."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
    "/api/friend-invitations/{friendInvitationId}": {
      "delete": {
        "tags": [
          "FriendInvitations"
        ],
        "summary": "Delete friendship invitation",
        "description": "Route: PATCH api/friend-invitations/{friendInvitationId}/decline\r\nAuthorization: Required.\r\nThe operation changes the status of the invitation to \"Delete\" and may be idempotent depending on domain rules.",
        "parameters": [
          {
            "name": "friendInvitationId",
            "in": "path",
            "description": "The identifier of the friend invitation to decline.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The invitation has been deleted."
          },
          "400": {
            "description": "Invalid request or the invitation cannot be deleted due to its current state."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The user is not authorized to delete this invitation."
          },
          "404": {
            "description": "Invitation not found."
          },
          "409": {
            "description": "A conflict occurred (e.g., invitation already accepted or otherwise finalized)."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
    "/api/friend-invitations/{friendInvitationId}/accept": {
      "patch": {
        "tags": [
          "FriendInvitations"
        ],
        "summary": "Accept friendship invitation",
        "description": "Route: POST api/friend-invitations/{friendInvitationId}/accept\r\nAuthorization: Required.\r\nSide effects: Creates a new friendship and consumes (removes or finalizes) the invitation.",
        "parameters": [
          {
            "name": "friendInvitationId",
            "in": "path",
            "description": "The identifier of the friend invitation to accept.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The invitation was accepted and a friendship was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInvitationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or the invitation cannot be accepted in its current state."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The user is not authorized to accept this invitation."
          },
          "404": {
            "description": "Invitation not found."
          },
          "409": {
            "description": "A conflict occurred (e.g., invitation already handled or friendship already exists)."
          },
          "410": {
            "description": "The invitation has expired."
          },
          "500": {
            "description": "Unexpected server error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInvitationResponse": {
        "type": "object",
        "properties": {
          "friendshipId": {
            "type": "string",
            "format": "uuid"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "SendFriendInvitationRequest": {
        "required": [
          "recipientId",
          "senderId"
        ],
        "type": "object",
        "properties": {
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "recipientId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "SendFriendInvitationResponse": {
        "type": "object",
        "properties": {
          "friendshipInvitationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}